+++ src/com/xiuye/util/cls/XType.java
@@ -12,8 +12,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
-import com.xiuye.util.code.XYClassLoader;
-
 /**
  * Type operator
  *
